ml2/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
catboost>=1.2
lightgbm>=4.0.0
xgboost>=2.0.0
PyYAML>=6.0.0
matplotlib>=3.7.0
shap>=0.45.0
jupyter>=1.0.0
